+++ fmtc/fmtc.go
@@ -13,7 +13,10 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"strconv"
+	"strings"
 
+	"pkg.re/essentialkaos/ek.v9/env"
 	"pkg.re/essentialkaos/ek.v9/strutil"
 )
 
@@ -211,6 +214,11 @@ func Bell() {
 	fmt.Printf(_CODE_BELL)
 }
 
+// Is256ColorsSupported return true if 256 colors is supported
+func Is256ColorsSupported() bool {
+	return strings.Contains(env.Get().GetS("TERM"), "256color")
+}
+
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // TPrintf remove the previous message (if printed) and print new message
@@ -242,45 +250,80 @@ func TPrintln(a ...interface{}) (int, error) {
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+// codebeat:disable[LOC,BLOCK_NESTING]
+
 func tag2ANSI(tag string, clean bool) string {
 	if clean {
 		return ""
 	}
 
-	var (
-		modificator = 0
-		charColor   = 39
-		bgColor     = 49
-		light       = false
-	)
+	if isExtendedColorTag(tag) {
+		return parseExtendedColor(tag)
+	}
+
+	light := strings.Contains(tag, "-")
+	reset := strings.Contains(tag, "!")
+
+	var chars string
 
 	for _, key := range tag {
 		code := codes[key]
 
+		switch {
+		case light && code == 37: // Light gray = Dark gray
+			chars += "90;"
+			continue
+		case light && code == 97: // Light gray = Dark gray
+			chars += "97;"
+			continue
+		}
+
 		switch key {
-		case '-':
-			light = true
-		case '!', '*', '^', '_', '~', '@':
-			modificator = code
-		case 'd', 'r', 'g', 'y', 'b', 'm', 'c', 's', 'w':
-			charColor = code
+		case '-', '!':
+			continue
+
+		case '*', '^', '_', '~', '@':
+			if reset {
+				chars += getResetCode(code)
+			} else {
+				chars += strconv.Itoa(code)
+			}
+
 		case 'D', 'R', 'G', 'Y', 'B', 'M', 'C', 'S', 'W':
-			bgColor = code
+			chars += strconv.Itoa(code)
+
+		case 'd', 'r', 'g', 'y', 'b', 'm', 'c', 's', 'w':
+			if light {
+				chars += strconv.Itoa(code + 60)
+			} else {
+				chars += strconv.Itoa(code)
+			}
 		}
+
+		chars += ";"
 	}
 
-	if light {
-		switch charColor {
-		case 97:
-			break
-		case 37:
-			charColor = 90
-		default:
-			charColor += 60
-		}
+	return fmt.Sprintf("\033[" + chars[:len(chars)-1] + "m")
+}
+
+// codebeat:enable[LOC,BLOCK_NESTING]
+
+func parseExtendedColor(tag string) string {
+	// Foreground
+	if strings.HasPrefix(tag, "#") {
+		return "\033[38;5;" + tag[1:] + "m"
 	}
 
-	return fmt.Sprintf("\033[%d;%d;%dm", modificator, charColor, bgColor)
+	// Background
+	return "\033[48;5;" + tag[1:] + "m"
+}
+
+func getResetCode(code int) string {
+	if code == codes['*'] {
+		code++
+	}
+
+	return "2" + strconv.Itoa(code)
 }
 
 func replaceColorTags(input, output *bytes.Buffer, clean bool) bool {
@@ -378,12 +421,35 @@ func getSymbols(symbol string, count int) string {
 }
 
 func isValidTag(tag string) bool {
+	if isValidExtendedTag(tag) {
+		return true
+	}
+
 	for _, r := range tag {
 		_, hasCode := codes[r]
 
 		if !hasCode {
 			return false
 		}
+	}
+
+	return true
+}
+
+func isExtendedColorTag(tag string) bool {
+	return strings.HasPrefix(tag, "#") || strings.HasPrefix(tag, "%")
+}
+
+func isValidExtendedTag(tag string) bool {
+	if !isExtendedColorTag(tag) {
+		return false
+	}
+
+	tag = strings.TrimLeft(tag, "#%")
+	color, err := strconv.Atoi(tag)
+
+	if err != nil || color < 0 || color > 256 {
+		return false
 	}
 
 	return true