+++ changelog.md
@@ -1,5 +1,15 @@
 ## Changelog
 
+### 9.19.0
+
+* `directio` Added sub-package `directio` for writing/reading data with using direct IO
+* `fmtc` 256 colors support with new tags (foreground: `{#000}`, background: `{%000}`)
+* `fmtc` Added method `Is256ColorsSupported` for checking support of 256 color output
+* `fmtc` Improved color tags syntax
+* `fmtc` Added tags for resetting modificators (e.g. `{!*}`)
+* `fmtc` Removed color tags overriding (i.e. now `{*}{r} == {r*}`)
+* `color` Added method `RGB2Term` for converting RGB colors to terminal color codes
+
 ### 9.18.1
 
 * `[system]` Fixed bug with extra new line symbol in user `Shell` field